feat: add manifest plugin to output manifest file & extract state initialization
